+++ test/fixture/svelte.config.js
@@ -27,6 +27,7 @@ export default {
 		adapter: adapter({
 			out: variant.out,
 			tracing: variant.tracing,
+			staticDotfiles: variant.staticDotfiles,
 			websocket: variant.handler
 				? { ...variant.websocket, handler: variant.handler }
 				: variant.websocket